convert to float64 from float32

+++ app/models/answers.go
@@ -11,13 +11,13 @@ type Answer struct {
 	QuizID           int       `json:"quiz_id"`
 	Correct          bool      `json:"correct"`
 	AnsweredOptionID int       `json:"answered_option_id"`
-	RemainedTime     float32   `json:"remained_time"`
+	RemainedTime     float64   `json:"remained_time"`
 	OnlineMatchID    int       `json:"online_match_id"`
 	CreatedAt        time.Time `json:"created_at"`
 	UpdatedAt        time.Time `json:"updated_at"`
 }
 
-func CalculateTotalRemainedTime(online_match_id int, user_id int) (total_time float32, err error) {
+func CalculateTotalRemainedTime(online_match_id int, user_id int) (total_time float64, err error) {
 	cmd := `select sum(remained_time) as total_time from answers where online_match_id = ? and user_id = ?`
 	rows, err := Db.Query(cmd, online_match_id, user_id)
 	if err != nil {
@@ -29,7 +29,7 @@ func CalculateTotalRemainedTime(online_match_id int, user_id int) (total_time fl
 		rows.Scan(&total_remained_time)
 	}
 
-	return float32(total_remained_time), err
+	return total_remained_time, err
 }
 
 func (a *Answer) CreateAnswer() (err error) {
