+++ app/models/websocket.go
@@ -7,15 +7,14 @@ import (
 type WsUser struct {
 	ID           string   `json:"id"`
 	Name         string   `json:"name"`
-	RemainedTime *float32 `json:"remained_time"`
+	RemainedTime *float64 `json:"remained_time"`
 	Icon         string   `json:"icon"`
 }
 
 type WsResponse struct {
-	Action                   string   `json:"action"`
-	// JoinedOnlineMatchUserIDs []string `json:"joined_onine_match_user_ids"`
-	Users                    []WsUser `json:"users"`
-	OnlineMatchID            int      `json:"online_match_id"`
+	Action        string   `json:"action"`
+	Users         []WsUser `json:"users"`
+	OnlineMatchID int      `json:"online_match_id"`
 }
 
 type WsConnection struct {
@@ -26,7 +25,7 @@ type WsRequest struct {
 	Action        string       `json:"action"`
 	UserID        string       `json:"user_id"`
 	UserName      string       `json:"user_name"`
-	RemainedTime  float32      `json:"remained_time"`
+	RemainedTime  float64      `json:"remained_time"`
 	UserIcon      string       `json:"user_icon"`
 	OnlineMatchID int          `json:"online_match_id"`
 	Conn          WsConnection `json:"-"`
